+++ packages/my-components/src/DatePicker/DatePickerInput/DatePickerInput.module.scss
