+++ packages/my-components/src/DatePicker/DatePickerInput/index.ts
@@ -1,0 +1,1 @@
+export { default } from './DatePickerInput';
