--- packages/my-components/src/DatePicker/DatePickerInputs/index.ts
@@ -1,1 +1,0 @@
-export { default } from './DatePickerInputs';
